ClampByte(Round(...)) performance fix "ClampByte(Round(...))" calls the ClampByte(double) function because Round() returns an Int64 which is too large for ClampByte(integer)). Casting the Round() return value to an Integer fix the "Double=>Int64=>Double=>Byte" conversion.

+++ source/Img32.SVG.Reader.pas
@@ -2136,7 +2136,7 @@ var
 begin
   if not GetSrcAndDst or not Assigned(values) then Exit;
   for i := 0 to 19 do
-    colorMatrix[i] := ClampByte(Round(values[i]*255));
+    colorMatrix[i] := ClampByte(Integer(Round(values[i]*255)));
 
   dx1 := srcImg.Width - RectWidth(srcRec);
   dx2 := dstImg.Width - RectWidth(dstRec);
